requirements.txt
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.1
certifi==2019.3.9
chardet==3.0.4
Django==1.8.2
django-haystack==2.6.1
django-mdeditor==0.1.13
django-redis==3.8.3
django-redis-sessions==0.5.6
django-tinymce==2.6.0
docopt==0.6.2
fdfs-client-py==1.2.6
idna==2.8
itsdangerous==1.1.0
jieba==0.39
kombu==4.4.0
mutagen==1.42.0
Pillow==5.4.1
pipreqs==0.4.9
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.0
pytz==2018.9
redis==2.10.6
requests==2.21.0
urllib3==1.24.1
uWSGI==2.0.18
vine==1.2.0
Whoosh==2.7.4
yarg==0.1.9
